+++ src/Data/Geo/Jord/Spherical/GreatCircle.hs
@@ -1,19 +1,17 @@
 -- |
--- Module:      Data.Geo.Jord.GreatCircle
+-- Module:      Data.Geo.Jord.Spherical.GreatCircle
 -- Copyright:   (c) 2018 Cedric Liegeois
 -- License:     BSD3
 -- Maintainer:  Cedric Liegeois <ofmooseandmen@yahoo.fr>
 -- Stability:   experimental
 -- Portability: portable
 --
--- Types and functions for working with <https://en.wikipedia.org/wiki/Great_circle Great Circle>.
+-- Types and functions for working with <https://en.wikipedia.org/wiki/Great_circle Great Circle> assuming a spherical earth model.
 --
 -- All functions are implemented using the vector-based approached described in
 -- <http://www.navlab.net/Publications/A_Nonsingular_Horizontal_Position_Representation.pdf Gade, K. (2010). A Non-singular Horizontal Position Representation>
 --
--- This module assumes a spherical earth.
---
-module Data.Geo.Jord.GreatCircle
+module Data.Geo.Jord.Spherical.GreatCircle
     (
     -- * The 'GreatCircle' type
       GreatCircle
@@ -32,11 +30,12 @@ module Data.Geo.Jord.GreatCircle
 import Control.Monad.Fail
 import Data.Geo.Jord.Angle
 import Data.Geo.Jord.Ellipsoid
-import Data.Geo.Jord.HorizontalPosition
 import Data.Geo.Jord.LatLong
 import Data.Geo.Jord.Length
+import Data.Geo.Jord.NVector
+import Data.Geo.Jord.Position (Geodetic2D(..), northPole)
 import Data.Geo.Jord.Quantity
-import Data.Geo.Jord.Vector3d
+import Data.Geo.Jord.Spherical.Geodetics
 import Data.Maybe (fromMaybe)
 import Prelude hiding (fail)
 
@@ -50,24 +49,24 @@ import Prelude hiding (fail)
 -- See 'greatCircle', 'greatCircleE', 'greatCircleF' or 'greatCircleBearing' constructors.
 --
 data GreatCircle = GreatCircle
-    { normal :: Vector3d
+    { normal :: NVector
     , dscr :: String
     } deriving (Eq)
 
 instance Show GreatCircle where
     show = dscr
 
--- | 'GreateCircle' passing by both given 'HorizontalPosition's. 'error's if given positions are
+-- | 'GreatCircle' passing by both given horizontal positions. 'error's if given positions are
 -- equal or antipodal.
-greatCircle :: (Eq a, HorizontalPosition a, Show a) => a -> a -> GreatCircle
+greatCircle :: (Eq a, Geodetic2D a, Show a) => a -> a -> GreatCircle
 greatCircle p1 p2 =
     fromMaybe
         (error (show p1 ++ " and " ++ show p2 ++ " do not define a unique Great Circle"))
         (greatCircleF p1 p2)
 
--- | 'GreateCircle' passing by both given 'Position's. A 'Left' indicates that given positions are
+-- | 'GreatCircle' passing by both given horizontal positions. A 'Left' indicates that given positions are
 -- equal or antipodal.
-greatCircleE :: (HorizontalPosition a) => a -> a -> Either String GreatCircle
+greatCircleE :: (Geodetic2D a) => a -> a -> Either String GreatCircle
 greatCircleE p1 p2
     | p1 == p2 = Left "Invalid Great Circle: positions are equal"
     | p1 == antipode p2 = Left "Invalid Great Circle: positions are antipodal"
@@ -81,9 +80,9 @@ greatCircleE p1 p2
     v1 = toNVector p1
     v2 = toNVector p2
 
--- | 'GreateCircle' passing by both given 'Position's. 'fail's if given positions are
+-- | 'GreatCircle' passing by both given horizontal positions. 'fail's if given positions are
 -- equal or antipodal.
-greatCircleF :: (MonadFail m, HorizontalPosition a) => a -> a -> m GreatCircle
+greatCircleF :: (MonadFail m, Geodetic2D a) => a -> a -> m GreatCircle
 greatCircleF p1 p2 =
     case e of
         Left err -> fail err
@@ -91,8 +90,8 @@ greatCircleF p1 p2 =
   where
     e = greatCircleE p1 p2
 
--- | 'GreatCircle' passing by the given 'Position' and heading on given bearing.
-greatCircleBearing :: (HorizontalPosition a) => a -> Angle -> GreatCircle
+-- | 'GreatCircle' passing by the given horizontal position and heading on given bearing.
+greatCircleBearing :: (Geodetic2D a) => a -> Angle -> GreatCircle
 greatCircleBearing p b =
     GreatCircle
         (sub n' e')
@@ -105,10 +104,10 @@ greatCircleBearing p b =
     n' = scale n (sin' b / norm n)
 
 -- | 'crossTrackDistance'' using the mean radius of the WGS84 reference ellipsoid.
-crossTrackDistance :: (HorizontalPosition a) => a -> GreatCircle -> Length
+crossTrackDistance :: (Geodetic2D a) => a -> GreatCircle -> Length
 crossTrackDistance p gc = crossTrackDistance' p gc (meanRadius wgs84)
 
--- | Signed distance from given 'Position' to given 'GreatCircle'.
+-- | @crossTrackDistance' p gc@ computes the signed distance horizontal position @p@ to great circle @gc@.
 -- Returns a negative 'Length' if position if left of great circle,
 -- positive 'Length' if position if right of great circle; the orientation of the
 -- great circle is therefore important:
@@ -118,14 +117,14 @@ crossTrackDistance p gc = crossTrackDistance' p gc (meanRadius wgs84)
 --     let gc2 = greatCircle (latLongDecimal 52 1) (latLongDecimal 51 0)
 --     crossTrackDistance p gc1 == (- crossTrackDistance p gc2)
 -- @
-crossTrackDistance' :: (HorizontalPosition a) => a -> GreatCircle -> Length -> Length
+crossTrackDistance' :: (Geodetic2D a) => a -> GreatCircle -> Length -> Length
 crossTrackDistance' p gc =
     arcLength (sub (angularDistance (normal gc) (toNVector p) Nothing) (decimalDegrees 90))
 
 -- | Computes the intersections between the two given 'GreatCircle's.
 -- Two 'GreatCircle's intersect exactly twice unless there are equal (regardless of orientation),
 -- in which case 'Nothing' is returned.
-intersections :: (HorizontalPosition a) => GreatCircle -> GreatCircle -> Maybe (a, a)
+intersections :: (Geodetic2D a) => GreatCircle -> GreatCircle -> Maybe (a, a)
 intersections gc1 gc2
     | norm i == 0.0 = Nothing
     | otherwise