+++ ChangeLog.md
@@ -1,7 +1,7 @@
 ### 0.3.0.0
 
-- Position -> HorizontalPosition
-- NVector -> Vector3d
+- Added ellipsoid, ECEF positions
+- Added NEDVector
 
 ### 0.2.0.0